requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
chromadb==0.5.5
google-generativeai==0.8.2
pypdf==4.3.1
python-dotenv==1.0.1
langgraph
langchain-google-genai
langchain-core
langchain
Pillow==10.4.0
faiss-cpu
